Module rfyl::rfyl

Structs

DiceRoll

DiceRolls

Functions

parse_into_infix
Returns a Vector of Strings with each element containing a token or an operator in bracketed infix format.

parse_into_rpn
Returns a Vector of Strings with each element containing a token or an operator in postfix (rpn) format.

roll
Returns a DiceRolls object based on the provided formula.

solve_rpn_formula
Returns an i32 as the result of a postfix (rpn) formula.