Function rfyl::rfyl::parse_into_rpn

pub fn parse_into_rpn(input_formula: &str) -> Vec<String>

Returns a Vector of Strings with each element containing a token or an operator in postfix (rpn) format.

Arguments

input_formula - A string that provides the notation to work off.

Example values

3 + 4 * 6 -> ["3", "4", "6", "*", "+"]

2d4 + d6 + d4 -> ["2d4", "d6", "d4", "+", "+"]

xv * (ab + dc) -> ["xv", "ab", "dc", "+", "*"]