Struct rfyl::rfyl::DiceRolls

pub struct DiceRolls { /* fields omitted */ }

Methods

impl DiceRolls

fn get_result(&self) -> i32

Returns an i32 as the result of the formula including any calculational operators.

fn get_sum_of_rolls(&self) -> i32

Returns an i32 as the simple sum of all rolls.

fn get_rolls_string(&self) -> String

Returns a formatted String showing the dice and the rolled results.

fn get_formula_string_as_rpn(&self) -> String

Returns a postfix formatted String showing the formula.

fn get_formula_string_as_infix(&self) -> String

Returns an infix formatted String showing the formula.

fn get_rolls_formula_string_as_rpn(&self) -> String

Returns a postfix formatted String showing the formula withthe original dice notation instead of the rolled result.

fn get_rolls_formula_string_as_infix(&self) -> String

Returns a infix formatted String showing the formula withthe original dice notation instead of the rolled result.